requirements.txt
appdirs==1.4.4
argcomplete==3.6.3
blinker==1.9.0
click==8.3.1
Flask==3.1.3
flask-cors==5.0.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.1.1
python-barcode==0.16.1
python-escpos==3.1
pyusb==1.3.1
PyYAML==6.0.3
qrcode==8.2
six==1.17.0
requests>=2.31
Werkzeug==3.1.6
zeroconf==0.146.0
pytest>=7.0
